requirements.txt
numpy>=1.24.0
librosa>=0.10.1
soundfile>=0.12.1
openai-whisper>=20231117
scikit-learn>=1.3.0
torch>=2.0.0  # whisper的依赖 